Add files via upload
First of all, I understood the function of size() method (yeah, I really should have read the about in Java Doc kek), and because of that, I created a method to do the lists' sizes in an auxiliary class: ListLimit(); 
I made a few new things: created the aux. class ListLimit inside the listWord method; made the lists may have different int values; and count the total of lines, columns, and Strings generated.